Refactor streaming to simplify for logging change (#28056)
